+++ sktime/benchmarking/experiments.py
@@ -14,13 +14,44 @@ import numpy as np
 from sklearn import preprocessing
 
 
+from sktime.utils.data_io import write_results_to_uea_format
+from sktime.utils.data_io import load_from_tsfile_to_dataframe as load_ts
+from sktime.utils.sampling import stratified_resample
 from sktime.clustering import (
     TimeSeriesKMeans,
     TimeSeriesKMedoids,
 )
-from sktime.utils.data_io import write_results_to_uea_format
-from sktime.utils.data_io import load_from_tsfile_to_dataframe as load_ts
-from sktime.utils.sampling import stratified_resample
+from sktime.classification.dictionary_based import (
+    BOSSEnsemble,
+    ContractableBOSS,
+    TemporalDictionaryEnsemble,
+    WEASEL,
+    MUSE,
+)
+from sktime.classification.distance_based import (
+    ElasticEnsemble,
+    ProximityForest,
+    ProximityTree,
+    ProximityStump,
+    KNeighborsTimeSeriesClassifier,
+    ShapeDTW,
+)
+from sktime.classification.feature_based import (
+    Catch22Classifier,
+    MatrixProfileClassifier,
+    SignatureClassifier,
+    TSFreshClassifier,
+)
+from sktime.classification.interval_based import (
+    TimeSeriesForestClassifier,
+    RandomIntervalSpectralForest,
+    SupervisedTimeSeriesForest,
+)
+from sktime.classification.interval_based import CanonicalIntervalForest, DrCIF
+from sktime.classification.kernel_based import ROCKETClassifier, Arsenal
+from sktime.classification.shapelet_based import MrSEQLClassifier
+from sktime.classification.shapelet_based import ShapeletTransformClassifier
+from sktime.classification.hybrid import HIVECOTEV1
 
 
 def run_clustering_experiment(
@@ -228,6 +259,8 @@ def load_and_run_clustering_experiment(
     )
 
 
+
+
 def set_clusterer(cls, resample_id=None):
     """Construct a clusterer.
 
@@ -267,3 +300,99 @@ def set_clusterer(cls, resample_id=None):
 
     else:
         raise Exception("UNKNOWN CLUSTERER")
+
+
+
+def set_classifier(cls, resampleId=None):
+    """Construct a classifier.
+
+    Basic way of creating the classifier to build using the default settings. This
+    set up is to help with batch jobs for multiple problems to facilitate easy
+    reproducability for use with load_and_run_classification_experiment. You can pass a
+    classifier object instead to run_classification_experiment.
+
+    Parameters
+    ----------
+    cls: String indicating which classifier you want
+    resampleId: classifier random seed
+
+    Return
+    ------
+    A BaseClassifier.
+    """
+    name = cls.lower()
+    # Dictionary based
+    if name == "boss" or name == "bossensemble":
+        return BOSSEnsemble(random_state=resampleId)
+    elif name == "cboss" or name == "contractableboss":
+        return ContractableBOSS(random_state=resampleId)
+    elif name == "tde" or name == "temporaldictionaryensemble":
+        return TemporalDictionaryEnsemble(random_state=resampleId)
+    elif name == "weasel":
+        return WEASEL(random_state=resampleId)
+    elif name == "muse":
+        return MUSE(random_state=resampleId)
+    # Distance based
+    elif name == "pf" or name == "proximityforest":
+        return ProximityForest(random_state=resampleId)
+    elif name == "pt" or name == "proximitytree":
+        return ProximityTree(random_state=resampleId)
+    elif name == "ps" or name == "proximityStump":
+        return ProximityStump(random_state=resampleId)
+    elif name == "dtwcv" or name == "kneighborstimeseriesclassifier":
+        return KNeighborsTimeSeriesClassifier(distance="dtwcv")
+    elif name == "dtw" or name == "1nn-dtw":
+        return KNeighborsTimeSeriesClassifier(distance="dtw")
+    elif name == "msm" or name == "1nn-msm":
+        return KNeighborsTimeSeriesClassifier(distance="msm")
+    elif name == "ee" or name == "elasticensemble":
+        return ElasticEnsemble()
+    elif name == "shapedtw":
+        return ShapeDTW()
+    # Feature based
+    elif name == "catch22":
+        return Catch22Classifier(
+            random_state=resampleId, estimator=RandomForestClassifier(n_estimators=500)
+        )
+    elif name == "matrixprofile":
+        return MatrixProfileClassifier(random_state=resampleId)
+    elif name == "signature":
+        return SignatureClassifier(
+            random_state=resampleId, classifier=RandomForestClassifier(n_estimators=500)
+        )
+    elif name == "tsfresh":
+        return TSFreshClassifier(
+            random_state=resampleId, estimator=RandomForestClassifier(n_estimators=500)
+        )
+    elif name == "tsfresh-r":
+        return TSFreshClassifier(
+            random_state=resampleId,
+            estimator=RandomForestClassifier(n_estimators=500),
+            relevant_feature_extractor=True,
+        )
+    # Hybrid
+    elif name == "hivecotev1":
+        return HIVECOTEV1(random_state=resampleId)
+    # Interval based
+    elif name == "rise" or name == "randomintervalspectralforest":
+        return RandomIntervalSpectralForest(random_state=resampleId, n_estimators=500)
+    elif name == "tsf" or name == "timeseriesforestclassifier":
+        return TimeSeriesForestClassifier(random_state=resampleId, n_estimators=500)
+    elif name == "cif" or name == "canonicalintervalforest":
+        return CanonicalIntervalForest(random_state=resampleId, n_estimators=500)
+    elif name == "stsf":
+        return SupervisedTimeSeriesForest(random_state=resampleId, n_estimators=500)
+    elif name == "drcif":
+        return DrCIF(random_state=resampleId, n_estimators=500)
+    # Kernel based
+    elif name == "rocket":
+        return ROCKETClassifier(random_state=resampleId)
+    elif name == "arsenal":
+        return Arsenal(random_state=resampleId)
+    # Shapelet based
+    elif name == "stc" or name == "shapelettransformclassifier":
+        return ShapeletTransformClassifier(random_state=resampleId, n_estimators=500)
+    elif name == "mrseql" or name == "mrseqlclassifier":
+        return MrSEQLClassifier(seql_mode="fs", symrep=["sax", "sfa"])
+    else:
+        raise Exception("UNKNOWN CLASSIFIER")
